move wheel logic to own class
